+++ custom_components/series_complete/const.py
@@ -2,6 +2,6 @@
 
 DOMAIN = "series_complete"
 
-DEFAULT_HOST = "localhost"
+DEFAULT_HOST = "e81ba94f-series-complete-plex"
 DEFAULT_PORT = 3000
 DEFAULT_SCAN_INTERVAL = 1800  # 30 minutes